refactor: consolidate markdown rendering pipeline and upgrade to v0.1.1
- Merge ast CSS module into html module, simplify CSS engine selectors

- Rewrite HTML parser/styling into unified html::styled::html_to_styled_nodes

- Expand block-level generator (tables, lists, images)

- Fix vello_cpu 0.1.0 API breakage (render_to_pixmap -> render)

- Fix clippy map-values iteration warnings in test helpers

- Bump version to 0.1.1

+++ tests/stage2_generator/layout.rs
@@ -12,8 +12,8 @@ fn get_element_bounds(elem: &VisualElement) -> (f64, f64, f64, f64) {
         VisualElement::Image { position, size, .. } => (
             position.x,
             position.y,
-            position.x + size.x as f64,
-            position.y + size.y as f64,
+            position.x + size.x,
+            position.y + size.y,
         ),
         _ => (0.0, 0.0, 0.0, 0.0),
     }
